==> Code/00_CleanMediaRatings.ipynb <==
import os 
import pandas as pd
# Set the working directory
os.chdir("/Users/dorianquelle/Desktop/zurich/TwitterViews/Code")

# Read in the data
data = pd.read_csv("../Mediadatabases/Clean_MediaRatingTwitter.csv",index_col=0)
# Remove Rows which are 1) Empty, 2) NA, 3) Not a Twitter Handle
data = data[(data["Twitter.Handle"].notna())]
data = data.reset_index(drop=True)
data = data.loc[:, ["Name.of.Outlet", "Political.Leaning","Twitter.Handle"]]
# Count number of @ per row
assert set(data["Twitter.Handle"].str.count("@")) == {0}, "At least one row has more or less than one @"
data = data[~data["Twitter.Handle"].duplicated()]
# Shuffle the data
data = data.sample(frac=1).reset_index(drop=True)
# Export to CSV
data = data[~data["Name.of.Outlet"].str.contains("bama")]
too_big = ['CNN','HuffPost','ABC News','Associated Press','The Guardian','The New York Times News','TIME','The Washington Post','Forbes','Reuters','The Wall Street Journal News','Fox News','barackobama.com','theguardian.com','huffingtonpost.com','vogue.com','mtv.com','wired.com','economist.com','nytimes.com','washingtonpost.com','billboard.com','theonion.com','reuters.com','time.com','techcrunch.com','cnn.com','eonline.com','apnews.com','abcnews.go.com','wsj.com','blogs.wsj.com','forbes.com','nfl.com','espn.com','foxnews.com']
small_data = data[~data["Name.of.Outlet"].isin(too_big)]

data.to_csv("../Mediadatabases/Clean_MediaRatingTwitter.csv", index=False)
small_data.to_csv("../Mediadatabases/small_Clean_MediaRatingTwitter.csv", index=False)

state_media = pd.read_excel("../Mediadatabases/StateMediaTwitter.xlsx")
# Parse Twitter Link to get Twitter Handle
state_media["Twitter Handle"] = state_media["Twitter Link"].str.split("/").str[-1]
# Remove Rows which are 1) Empty, 2) NA, 3) Not a Twitter Handle
state_media = state_media[(state_media["Twitter Handle"].notna()) & (state_media["Twitter Handle"] != "")]
state_media = state_media.reset_index(drop=True)
state_media = state_media.loc[:, ["Name", "Country","Twitter Handle"]]
state_media.to_csv("../Mediadatabases/Clean_StateMediaTwitter.csv", index=False)
